rescue 404 on missing key
Consul, as a HTTP API key value store handles missing key properly by
returning a 404 code.

Faraday itself handles that by raising an exception and it's then down
to the application code to do something about it.

Usually key value stores libraries return a nil or empty string value
for missing keys. A nil value is more manageable in Ruby.

The patch adds a rescue on the Faraday::ResourceNotFound exception
around the @conn.get. This rescue set the @raw attribute to a simple
hash with a nil Value.

The return_value method is then still called to use the rest of the
chain as usual.

A spec for the get on a missing key has been added with some
refactoring. A similar spec can be added in the ACL block too.