add proper parsing of recurse responses
Consul kv/ can handle a recurse option on the GET v1/kv/:key endpoint.
In that case Consul returns an array of entries matching the key.

It's an handy way to namespace or group keys based on a prefix of sorts.

As it is Diplomat would only returns one entry in the response even if
the recurse option is passed with the key.

This patch allows Diplomat to return an Array of {key, value} entries
that were returned by consul if there is more than 1 entry in the
response.

As this will not happen if the ?recurse option is not passed in the key
(Diplomat.kv.get("some?recurse")) I would consider it a non breaking
change but an extension of the current behavior.

Another solution would be to add a get_recurse method reusing the same
principle added in the return_value method.

See added spec for behavior too.