Changed Kv.get to always return an array for recursive calls.
When executing a recursive Kv.get, and there is only one item to
return, Diplomat was returning a string instead of an array of
key/value pairs. This means the caller has no idea what the key
is for the value returned.

To fix the problem, we've added a call to RestClient.return_value
that indicates that it whether this is a recurse get.

I added a spec to test the above behavoir.

One worrisome behavoir remains: if you call Kv.get() with :return
and there is no response, I'm not sure what is the proper response.
For recursive searches, I pretty sure we're still just returning an
empty string. And I think that recursive searches should always
return an array.