Don't strip away Consul metadata if we return a single value
Fixes an inconsistency outlined in WeAreFarmGeek/diplomat#39 where a KV `get`
call returns an array if >1 value is returned, but if one value is returned it
strips away the metadata and hands back the raw value.

While this is a friendly thing to do in some cases, it creates a very annoying
inconsistency when doing a recursive `get` which leaves out the metadata
(most importantly the key of the returned value) as well as requiring the
consumer to check whether the return value of `get` is a String or Array.