Changelog for release 2.1.0
This release cleanup a lot the existing APIs while preserving ascending compatibility.
It will avoid relying on side effects to configure diplomat and allow to override most
configuration options per API call as implemented in https://github.com/WeAreFarmGeek/diplomat/pull/179.
It is now easy to use one instance of the lib and to perform several calls with different tokens and/or
consistency options for instance.